Move AccountDetailsDisplay to its own file
