+++ ui/components/multichain/account-details/account-details.js
@@ -1,10 +1,9 @@
-import React, { useCallback, useContext, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import Popover from '../../ui/popover/popover.component';
 import {
   setAccountDetailsAddress,
-  setAccountLabel,
   clearAccountDetails,
   exportAccount,
   hideWarning,
@@ -13,7 +12,6 @@ import {
   AvatarAccount,
   AvatarAccountSize,
   AvatarAccountVariant,
-  BUTTON_SECONDARY_SIZES,
   BannerAlert,
   ButtonIcon,
   ButtonPrimary,
@@ -24,19 +22,7 @@ import {
   TextField,
 } from '../../component-library';
 import Box from '../../ui/box/box';
-import EditableLabel from '../../ui/editable-label/editable-label';
-import {
-  getMetaMaskAccountsOrdered,
-  getMetaMaskKeyrings,
-} from '../../../selectors';
-import QrView from '../../ui/qr-code';
-import { isHardwareKeyring } from '../../../helpers/utils/hardware';
-import { MetaMetricsContext } from '../../../contexts/metametrics';
-import {
-  MetaMetricsEventCategory,
-  MetaMetricsEventKeyType,
-  MetaMetricsEventName,
-} from '../../../../shared/constants/metametrics';
+import { getMetaMaskAccountsOrdered } from '../../../selectors';
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import {
   AlignItems,
@@ -53,17 +39,14 @@ import {
 } from '../../../helpers/constants/design-system';
 import { AddressCopyButton } from '../address-copy-button';
 import { useCopyToClipboard } from '../../../hooks/useCopyToClipboard';
+import { AccountDetailsDisplay } from './account-details-display';
 
 export const AccountDetails = ({ address }) => {
   const dispatch = useDispatch();
   const t = useI18nContext();
-  const trackEvent = useContext(MetaMetricsContext);
   const useBlockie = useSelector((state) => state.metamask.useBlockie);
-  const keyrings = useSelector(getMetaMaskKeyrings);
   const accounts = useSelector(getMetaMaskAccountsOrdered);
   const { name } = accounts.find((account) => account.address === address);
-  const keyring = keyrings.find((kr) => kr.accounts.includes(address));
-  const exportPrivateKeyFeatureEnabled = !isHardwareKeyring(keyring?.type);
 
   const [attemptingExport, setAttemptingExport] = useState(false);
   const [password, setPassword] = useState('');
@@ -138,42 +121,14 @@ export const AccountDetails = ({ address }) => {
       }
       onClose={onClose}
     >
-      <Box
-        display={DISPLAY.FLEX}
-        alignItems={AlignItems.center}
-        flexDirection={FLEX_DIRECTION.COLUMN}
-      >
-        {attemptingExport === false ? (
-          <>
-            <EditableLabel
-              defaultValue={name}
-              onSubmit={(label) => dispatch(setAccountLabel(address, label))}
-              accounts={accounts}
-            />
-            <QrView Qr={{ data: address }} />
-            {exportPrivateKeyFeatureEnabled ? (
-              <ButtonSecondary
-                block
-                size={BUTTON_SECONDARY_SIZES.LG}
-                variant={TextVariant.bodyMd}
-                onClick={() => {
-                  trackEvent({
-                    category: MetaMetricsEventCategory.Accounts,
-                    event: MetaMetricsEventName.KeyExportSelected,
-                    properties: {
-                      key_type: MetaMetricsEventKeyType.Pkey,
-                      location: 'Account Details Modal',
-                    },
-                  });
-                  setAttemptingExport(true);
-                }}
-              >
-                {t('showPrivateKey')}
-              </ButtonSecondary>
-            ) : null}
-          </>
-        ) : null}
-      </Box>
+      {attemptingExport === false ? (
+        <AccountDetailsDisplay
+          accounts={accounts}
+          accountName={name}
+          address={address}
+          onExportClick={() => setAttemptingExport(true)}
+        />
+      ) : null}
       {attemptingExport ? (
         <>
           <Box
